Samuel

+++ contracts/Puppy_Attack.sol
@@ -7,13 +7,13 @@ contract Puppy_Attack {
     using SafeMath for uint;
 
     // Constants
-    uint reviveRate = 0.01 ether;
-    uint levelUpFee = 0.005 ether;
+    uint256 reviveRate = 0.01 ether;
+    uint256 levelUpFee = 0.005 ether;
 
     struct Puppy {
-        uint id;
+        uint256 id;
         string name;
-        uint level;
+        uint256 level;
         address owner;
     }
 
@@ -39,13 +39,24 @@ contract Puppy_Attack {
     }
 
     function _create(string memory puppyName) public {
-        Puppy memory newPuppy = Puppy(allPuppies.length, puppyName, 5, msg.sender);
+        Puppy memory newPuppy =
+            Puppy(allPuppies.length, puppyName, 5, msg.sender);
         allPuppies.push(newPuppy);
     }
 
-    function _levelUp(uint puppyId) public payable restrictedToOwner(puppyId) {
-        require(allPuppies[puppyId].level > 0, "Your puppy needs to be revived before you can level it up.");
-        require(msg.value == levelUpFee, "Please pay the exact level up fee of 0.005 ether.");
+    function _levelUp(uint256 puppyId)
+        public
+        payable
+        restrictedToOwner(puppyId)
+    {
+        require(
+            allPuppies[puppyId].level > 0,
+            "Your puppy needs to be revived before you can level it up."
+        );
+        require(
+            msg.value == levelUpFee,
+            "Please pay the exact level up fee of 0.005 ether."
+        );
         allPuppies[puppyId].level = allPuppies[puppyId].level.add(1);
     }
 
@@ -63,9 +74,13 @@ contract Puppy_Attack {
         Puppy storage attackingPuppy = allPuppies[_attackingPuppyId];
         Puppy storage targetedPuppy = allPuppies[_targetedPuppyId];
 
-        uint rand = randMod(100);
-        if (attackingPuppy.level > targetedPuppy.level && rand > 70 || attackingPuppy.level == targetedPuppy.level && rand > 50 || attackingPuppy.level < targetedPuppy.level && rand > 30) {
-                emit FailAttack(_attackingPuppyId, _targetedPuppyId);
+        uint256 rand = randMod(100);
+        if (
+            (attackingPuppy.level > targetedPuppy.level && rand > 70) ||
+            (attackingPuppy.level == targetedPuppy.level && rand > 50) ||
+            (attackingPuppy.level < targetedPuppy.level && rand > 30)
+        ) {
+            emit FailAttack(_attackingPuppyId, _targetedPuppyId);
         } else {
             attackingPuppy.level = attackingPuppy.level.add(1);
             targetedPuppy.level = targetedPuppy.level.sub(1);
@@ -73,10 +88,12 @@ contract Puppy_Attack {
         }
     }
 
-    function randMod(uint _modulus) internal view returns(uint) {
-        uint randNonce = 0;
+    function randMod(uint256 _modulus) internal view returns (uint256) {
+        uint256 randNonce = 0;
         randNonce = randNonce.add(1);
-        return uint(keccak256(abi.encodePacked(now, msg.sender, randNonce))) % _modulus;
+        return
+            uint256(keccak256(abi.encodePacked(now, msg.sender, randNonce))) %
+            _modulus;
     }
 
     function revive(uint puppyId) public payable restrictedToOwner(puppyId) {
